refactor: update shared components
